requirements.txt
google-genai
python-dotenv
appwrite
mutagen